+++ packages/forms/sagas/index.js
@@ -1,9 +1,7 @@
 import { takeLatest } from 'redux-saga/effects';
-import { REQUEST_SUBMIT } from 'actions/types';
+import { REQUEST_SUBMIT } from '../actions/types';
 import watchRequestSubmit from './formSaga';
 
-export default function getSagas() {
-  return [
-    takeLatest(REQUEST_SUBMIT, watchRequestSubmit),
-  ];
-}
+export default () => ([
+  takeLatest(REQUEST_SUBMIT, watchRequestSubmit),
+]);